+++ Connectors/LinearConnector.cpp
@@ -1,6 +1,6 @@
 #include "LinearConnector.hpp"
 
-void	LinearConnector::setValue(int64_t AnalysisValue)
+void	LinearConnector::setValue(int16_t AnalysisValue)
 {
     value = AnalysisValue;
 }
@@ -9,6 +9,6 @@ LinearConnector::LinearConnector(){
     value = 255;
 }
 
-int32_t LinearConnector::getValue(){
+int16_t LinearConnector::getValue(){
             return value;
 }
