+++ include/AbstractAnalyzer.hpp
@@ -14,11 +14,11 @@ public:
     AbstractAnalyzer(AudioStream *as); 
     AbstractAnalyzer(const AbstractAnalyzer &a); // Copy Constructor
     void registerConnector(Connector* toInform);
-    int64_t getValue();
+    int16_t getValue();
     virtual void setValue(std::vector<int16_t> *PCMCodes) = 0; //This will be where each Analyzer does its magic.
     ~AbstractAnalyzer(); // Destructor
 protected:
-    int64_t value;
+    int16_t value;
     vector<Connector *> observers;
     AudioStream *audioStream;
